Module higher_kinded_lifetimes::nav_types

Types that can traverse the graph and cannot modify its structure. These types are backed by a read-only borrow of graph data, so multiple instances of them may coexist.

These traits may be thought of as fat pointers into some graph structure, augmented operations to navigate it.

Traits

Edge

InEdgeCollection

NavTypes

OutEdgeCollection

Vertex